refactored and renamed the SaveParametersButton and SavePythonButton to an abstract SaveFileButton component.
